+++ src/components/add.rs
@@ -1,5 +1,6 @@
 use crate::common::{
-    Component, Id, Input, OutputType, Ports, SignalSigned, SignalUnsigned, SignalValue, Simulator,
+    Component, Condition, Id, Input, OutputType, Ports, SignalSigned, SignalUnsigned, SignalValue,
+    Simulator,
 };
 use log::*;
 use serde::{Deserialize, Serialize};
@@ -31,21 +32,26 @@ impl Component for Add {
     }
 
     // propagate addition to output
-    fn clock(&self, simulator: &mut Simulator) {
+    fn clock(&self, simulator: &mut Simulator) -> Result<(), Condition> {
         // get input values
         let a_in = u32::try_from(simulator.get_input_value(&self.a_in));
         let b_in = u32::try_from(simulator.get_input_value(&self.b_in));
 
-        let (value, overflow) = match (&a_in, &b_in) {
+        let (value, overflow, res) = match (&a_in, &b_in) {
             (Ok(a), Ok(b)) => {
                 let (res, overflow) =
                     SignalSigned::overflowing_add(*a as SignalSigned, *b as SignalSigned);
                 (
                     (res as SignalUnsigned).into(),
                     (overflow as SignalUnsigned).into(),
+                    Ok(()),
                 )
             }
-            _ => (SignalValue::Unknown, SignalValue::Unknown),
+            _ => (
+                SignalValue::Unknown,
+                SignalValue::Unknown,
+                Err(Condition::Warning("Unknown".to_string())),
+            ),
         };
 
         trace!(
@@ -59,6 +65,7 @@ impl Component for Add {
         // set output
         simulator.set_out_value(&self.id, "out", value);
         simulator.set_out_value(&self.id, "overflow", overflow);
+        res
     }
 }