+++ src/components/probe_assert.rs
@@ -1,4 +1,4 @@
-use crate::common::{Component, Id, Input, OutputType, Ports, Signal, Simulator};
+use crate::common::{Component, Condition, Id, Input, OutputType, Ports, Signal, Simulator};
 use log::*;
 use serde::{Deserialize, Serialize};
 use std::rc::Rc;
@@ -24,14 +24,22 @@ impl Component for ProbeAssert {
         )
     }
 
-    fn clock(&self, simulator: &mut Simulator) {
+    fn clock(&self, simulator: &mut Simulator) -> Result<(), Condition> {
         trace!("-- cycle {} --", simulator.cycle);
+        let lhs = simulator.get_input_value(&self.input);
+        let rhs = self.values[simulator.cycle].get_value();
+
         // the assertion is checked only in test mode
         #[cfg(test)]
-        assert_eq!(
-            simulator.get_input_value(&self.input),
-            self.values[simulator.cycle].get_value()
-        );
+        assert_eq!(lhs, rhs);
+        if lhs == rhs {
+            Ok(())
+        } else {
+            Err(Condition::Assert(format!(
+                "assertion failed {:?} != {:?}",
+                lhs, rhs
+            )))
+        }
     }
 
     // notice we don't implement `un_clock` since the state is already kept in history